feat(rew): refactor biquad into its own mod
